+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTreeUI.java
@@ -17,10 +17,10 @@
 package com.formdev.flatlaf.ui;
 
 import static com.formdev.flatlaf.FlatClientProperties.*;
-
 import java.awt.Color;
 import java.awt.Component;
 import java.awt.Graphics;
+import java.awt.Graphics2D;
 import java.awt.Insets;
 import java.awt.Rectangle;
 import java.awt.event.MouseEvent;
@@ -36,6 +36,7 @@ import javax.swing.LookAndFeel;
 import javax.swing.SwingUtilities;
 import javax.swing.UIManager;
 import javax.swing.JTree.DropLocation;
+import javax.swing.event.TreeSelectionListener;
 import javax.swing.plaf.ComponentUI;
 import javax.swing.plaf.basic.BasicTreeUI;
 import javax.swing.tree.DefaultTreeCellRenderer;
@@ -96,6 +97,8 @@ import com.formdev.flatlaf.util.UIScale;
  * @uiDefault Tree.selectionForeground				Color
  * @uiDefault Tree.selectionInactiveBackground		Color
  * @uiDefault Tree.selectionInactiveForeground		Color
+ * @uiDefault Tree.selectionInsets					Insets
+ * @uiDefault Tree.selectionArc						int
  * @uiDefault Tree.wideSelection					boolean
  * @uiDefault Tree.showCellFocusIndicator			boolean
  *
@@ -132,6 +135,8 @@ public class FlatTreeUI
 	@Styleable protected Color selectionInactiveBackground;
 	@Styleable protected Color selectionInactiveForeground;
 	@Styleable protected Color selectionBorderColor;
+	/** @since 3 */ @Styleable protected Insets selectionInsets;
+	/** @since 3 */ @Styleable protected int selectionArc;
 	@Styleable protected boolean wideSelection;
 	@Styleable protected boolean showCellFocusIndicator;
 
@@ -175,6 +180,8 @@ public class FlatTreeUI
 		selectionInactiveBackground = UIManager.getColor( "Tree.selectionInactiveBackground" );
 		selectionInactiveForeground = UIManager.getColor( "Tree.selectionInactiveForeground" );
 		selectionBorderColor = UIManager.getColor( "Tree.selectionBorderColor" );
+		selectionInsets = UIManager.getInsets( "Tree.selectionInsets" );
+		selectionArc = UIManager.getInt( "Tree.selectionArc" );
 		wideSelection = UIManager.getBoolean( "Tree.wideSelection" );
 		showCellFocusIndicator = UIManager.getBoolean( "Tree.showCellFocusIndicator" );
 
@@ -296,6 +303,34 @@ public class FlatTreeUI
 	}
 
 	@Override
+	protected TreeSelectionListener createTreeSelectionListener() {
+		TreeSelectionListener superListener = super.createTreeSelectionListener();
+		return e -> {
+			superListener.valueChanged( e );
+
+			// for united rounded selection, repaint parts of the rows that adjoin to the changed rows
+			TreePath[] changedPaths;
+			if( useUnitedRoundedSelection() &&
+				tree.getSelectionCount() > 1 &&
+				(changedPaths = e.getPaths()) != null )
+			{
+				if( changedPaths.length > 4 ) {
+					// same is done in BasicTreeUI.Handler.valueChanged()
+					tree.repaint();
+				} else {
+					int arcHeight = (int) Math.ceil( UIScale.scale( (float) selectionArc ) );
+
+					for( TreePath path : changedPaths ) {
+						Rectangle r = getPathBounds( tree, path );
+						if( r != null )
+							tree.repaint( r.x, r.y - arcHeight, r.width, r.height + (arcHeight * 2) );
+					}
+				}
+			}
+		};
+	}
+
+	@Override
 	public Rectangle getPathBounds( JTree tree, TreePath path ) {
 		Rectangle bounds = super.getPathBounds( tree, path );
 
@@ -347,7 +382,7 @@ public class FlatTreeUI
 	}
 
 	/**
-	 * Same as super.paintRow(), but supports wide selection and uses
+	 * Similar to super.paintRow(), but supports wide selection and uses
 	 * inactive selection background/foreground if tree is not focused.
 	 */
 	@Override
@@ -428,7 +463,7 @@ public class FlatTreeUI
 				paintWideSelection( g, clipBounds, insets, bounds, path, row, isExpanded, hasBeenExpanded, isLeaf );
 			} else {
 				// non-wide selection
-				paintCellBackground( g, rendererComponent, bounds );
+				paintCellBackground( g, rendererComponent, bounds, row, true );
 			}
 
 			// this is actually not necessary because renderer should always set color
@@ -442,7 +477,7 @@ public class FlatTreeUI
 				if( bg != null && !bg.equals( defaultCellNonSelectionBackground ) ) {
 					Color oldColor = g.getColor();
 					g.setColor( bg );
-					paintCellBackground( g, rendererComponent, bounds );
+					paintCellBackground( g, rendererComponent, bounds, row, false );
 					g.setColor( oldColor );
 				}
 			}
@@ -497,7 +532,18 @@ public class FlatTreeUI
 	private void paintWideSelection( Graphics g, Rectangle clipBounds, Insets insets, Rectangle bounds,
 		TreePath path, int row, boolean isExpanded, boolean hasBeenExpanded, boolean isLeaf )
 	{
-		g.fillRect( 0, bounds.y, tree.getWidth(), bounds.height );
+		float arcTop, arcBottom;
+		arcTop = arcBottom = UIScale.scale( selectionArc / 2f );
+
+		if( useUnitedRoundedSelection() ) {
+			if( row > 0 && tree.isRowSelected( row - 1 ) )
+				arcTop = 0;
+			if( row < tree.getRowCount() - 1 && tree.isRowSelected( row + 1 ) )
+				arcBottom = 0;
+		}
+
+		FlatUIUtils.paintSelection( (Graphics2D) g, 0, bounds.y, tree.getWidth(), bounds.height,
+			UIScale.scale( selectionInsets ), arcTop, arcTop, arcBottom, arcBottom, 0 );
 
 		// paint expand/collapse icon
 		// (was already painted before, but painted over with wide selection)
@@ -507,7 +553,9 @@ public class FlatTreeUI
 		}
 	}
 
-	private void paintCellBackground( Graphics g, Component rendererComponent, Rectangle bounds ) {
+	private void paintCellBackground( Graphics g, Component rendererComponent, Rectangle bounds,
+		int row, boolean paintSelection )
+	{
 		int xOffset = 0;
 		int imageOffset = 0;
 
@@ -520,7 +568,32 @@ public class FlatTreeUI
 			xOffset = label.getComponentOrientation().isLeftToRight() ? imageOffset : 0;
 		}
 
-		g.fillRect( bounds.x + xOffset, bounds.y, bounds.width - imageOffset, bounds.height );
+		if( paintSelection ) {
+			float arcTopLeft, arcTopRight, arcBottomLeft, arcBottomRight;
+			arcTopLeft = arcTopRight = arcBottomLeft = arcBottomRight = UIScale.scale( selectionArc / 2f );
+
+			if( useUnitedRoundedSelection() ) {
+				if( row > 0 && tree.isRowSelected( row - 1 ) ) {
+					Rectangle r = getPathBounds( tree, tree.getPathForRow( row - 1 ) );
+					arcTopLeft = Math.min( arcTopLeft, r.x - bounds.x );
+					arcTopRight = Math.min( arcTopRight, (bounds.x + bounds.width) - (r.x + r.width) );
+				}
+				if( row < tree.getRowCount() - 1 && tree.isRowSelected( row + 1 ) ) {
+					Rectangle r = getPathBounds( tree, tree.getPathForRow( row + 1 ) );
+					arcBottomLeft = Math.min( arcBottomLeft, r.x - bounds.x );
+					arcBottomRight = Math.min( arcBottomRight, (bounds.x + bounds.width) - (r.x + r.width) );
+				}
+			}
+
+			FlatUIUtils.paintSelection( (Graphics2D) g, bounds.x + xOffset, bounds.y, bounds.width - imageOffset, bounds.height,
+				UIScale.scale( selectionInsets ), arcTopLeft, arcTopRight, arcBottomLeft, arcBottomRight, 0 );
+		} else
+			g.fillRect( bounds.x + xOffset, bounds.y, bounds.width - imageOffset, bounds.height );
+	}
+
+	private boolean useUnitedRoundedSelection() {
+		return selectionArc > 0 &&
+			(selectionInsets == null || (selectionInsets.top == 0 && selectionInsets.bottom == 0));
 	}
 
 	/**