+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyleableInfo.java
@@ -939,6 +939,8 @@ public class TestFlatStyleableInfo
 			"selectionInactiveBackground", Color.class,
 			"selectionInactiveForeground", Color.class,
 			"selectionBorderColor", Color.class,
+			"selectionInsets", Insets.class,
+			"selectionArc", int.class,
 			"wideSelection", boolean.class,
 			"showCellFocusIndicator", boolean.class,