+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyling.java
@@ -1145,6 +1145,8 @@ public class TestFlatStyling
 		ui.applyStyle( "selectionInactiveBackground: #fff" );
 		ui.applyStyle( "selectionInactiveForeground: #fff" );
 		ui.applyStyle( "selectionBorderColor: #fff" );
+		ui.applyStyle( "selectionInsets: 1,2,3,4" );
+		ui.applyStyle( "selectionArc: 8" );
 		ui.applyStyle( "wideSelection: true" );
 		ui.applyStyle( "showCellFocusIndicator: true" );